[REF] clipboard: Rename handlers files
A simple commit to change the name of the clipboard handlers' files
to avoid having once again a 'cell.ts', 'borders.ts', ... files
