+++ src/clipboard_handlers/chart_clipboard.ts
