+++ src/clipboard_handlers/data_validation_clipboard.ts
@@ -26,10 +26,10 @@ export class DataValidationClipboardHandler extends AbstractCellClipboardHandler
       return;
     }
 
-    const { rowsIndex, columnsIndex } = data;
+    const { rowsIndexes, columnsIndexes } = data;
     const sheetId = this.getters.getActiveSheetId();
-    const cellPositions = rowsIndex.map((row) =>
-      columnsIndex.map((col) => ({ col, row, sheetId }))
+    const cellPositions = rowsIndexes.map((row) =>
+      columnsIndexes.map((col) => ({ col, row, sheetId }))
     );
 
     return {
