+++ src/clipboard_handlers/filter_clipboard.ts
