+++ src/clipboard_handlers/image_clipboard.ts
