+++ src/clipboard_handlers/merge_clipboard.ts
