+++ src/clipboard_handlers/sheet_clipboard.ts
